refactor: keep normal path aliases
