+++ standalone/src/models/cart/cart/direct-discount/builders.spec.ts
@@ -1,4 +1,4 @@
-import { CartDiscountLineItemsTarget } from '@/models/cart-discount';
+import { CartDiscountLineItemsTarget } from '@/models/cart/cart-discount';
 import { DirectDiscountRest, DirectDiscountGraphql } from './index';
 
 describe('DirectDiscount Builder', () => {
