+++ standalone/src/models/order/generator.ts
@@ -1,6 +1,6 @@
 import { Generator, fake, oneOf, sequence } from '@/core';
-import { Cart, LineItem } from '@/models/cart';
-import { CartDiscount } from '@/models/cart-discount';
+import { Cart, LineItem } from '@/models/cart/cart';
+import { CartDiscount } from '@/models/cart/cart-discount';
 import { CentPrecisionMoney, ClientLogging, Address } from '@/models/commons';
 import * as CustomerGroup from '@/models/customer-group';
 import { createRelatedDates } from '@/utils';
