+++ standalone/src/models/staged-quote/generator.ts
@@ -1,6 +1,6 @@
 import { sequence, fake, Generator, oneOf } from '@/core';
 import { Company } from '@/models/business-unit';
-import { Cart } from '@/models/cart';
+import { Cart } from '@/models/cart/cart';
 import { ClientLogging } from '@/models/commons';
 import { Customer } from '@/models/customer';
 import { QuoteRequest } from '@/models/quote-request';
